+++ CHANGELOG.md
@@ -9,9 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Support for *easier secrets rotation* for JWT keys by adding new constructor `middlewares.NewAuthentication(...)`
+
 ### Changed
 
 ### Deprecated
+
+- Constructor `middlewares.NewAuth(...)` in favor of `middlewares.NewAuthentication(...)`
+    For 1:1 backwards compatibility use: `NewAuthentication(token, AuthWithRSAPublicKey(key), handlerFactory, options...)` instead of `NewAuth(token, key, handlerFactory, options...)`.
+    To support *easier secrets rotation* use: `NewAuthentication(token, AuthWithPublicKeyProvider(viperConfig), handlerFactory, options...)`
 
 ### Removed